requirements.txt
chainlit==1.3.0
langchain==0.3.3
langchain-core==0.3.12
langchain-community==0.3.2
langchain-ollama==0.2.0
langsmith==0.1.136
opencage==3.0.2
openai==1.9.0
pinecone==5.3.1
PyPDF2==3.0.1
python-dotenv==1.0.0
langchain-chroma==0.1.4